- Moving simulation date to the process simulation info logic module to avoid cyclic dependencies - updating the class javadoc comment
